+++ air-quality-ui/system_tests/e2e/ui/breadcrumb.spec.ts
@@ -1,0 +1,23 @@
+import { expect, test } from '../utils/fixtures'
+
+test('Mocked response breadcrumb', async ({
+  vairifySummaryPage,
+  vairifyCityPage,
+}) => {
+  await vairifySummaryPage.setupPage()
+
+  await vairifySummaryPage.clickButton('Kampala')
+  await expect(vairifyCityPage.textFinder('Cities')).toBeVisible()
+  await expect(vairifyCityPage.textFinder('Kampala')).toBeVisible()
+  await vairifySummaryPage.clickButton('Cities')
+
+  await vairifySummaryPage.clickButton('Abu Dhabi')
+  await expect(vairifyCityPage.textFinder('Cities')).toBeVisible()
+  await expect(vairifyCityPage.textFinder('Abu Dhabi')).toBeVisible()
+  await vairifySummaryPage.clickButton('Cities')
+
+  await vairifySummaryPage.clickButton('Zurich')
+  await expect(vairifyCityPage.textFinder('Cities')).toBeVisible()
+  await expect(vairifyCityPage.textFinder('Zurich')).toBeVisible()
+  await vairifySummaryPage.clickButton('Cities')
+})
